+++ src/main/java/com/miscitems/MiscItemsAndBlocks/Container/ContainerCharger.java
@@ -72,8 +72,8 @@ public class ContainerCharger  extends Container {
     public void addCraftingToCrafters(ICrafting par1ICrafting)
     {
         super.addCraftingToCrafters(par1ICrafting);
-        par1ICrafting.sendProgressBarUpdate(this, 0, this.tile.GetPower());
-        par1ICrafting.sendProgressBarUpdate(this, 1, this.tile.GetMaxPower());
+        par1ICrafting.sendProgressBarUpdate(this, 0, (int)this.tile.GetPower());
+        par1ICrafting.sendProgressBarUpdate(this, 1, (int)this.tile.GetMaxPower());
     }
 
     public void detectAndSendChanges()
@@ -86,18 +86,18 @@ public class ContainerCharger  extends Container {
 
             if (this.LastPower != this.tile.GetPower())
             {
-                icrafting.sendProgressBarUpdate(this, 0, this.tile.GetPower());
+                icrafting.sendProgressBarUpdate(this, 0, (int)this.tile.GetPower());
             }
             
             if (this.LastMaxPower != this.tile.GetMaxPower())
             {
-                icrafting.sendProgressBarUpdate(this, 1, this.tile.GetMaxPower());
+                icrafting.sendProgressBarUpdate(this, 1, (int)this.tile.GetMaxPower());
             }
             
         }
 
-        this.LastPower = this.tile.GetPower();
-        this.LastMaxPower = this.tile.GetMaxPower();
+        this.LastPower = (int)this.tile.GetPower();
+        this.LastMaxPower = (int)this.tile.GetMaxPower();
     }
 
     @SideOnly(Side.CLIENT)
