+++ src/main/java/com/miscitems/MiscItemsAndBlocks/Gui/GuiMiningChamber.java
@@ -48,14 +48,14 @@ public class GuiMiningChamber extends GuiContainer{
 	         int y = (this.height - this.ySize) / 2;
 	         drawTexturedModalRect(guiLeft, guiTop, 0, 0, xSize, ySize);
 
-	         int Power = this.tile.GetPower();
+	         int Power = (int)this.tile.GetPower();
 	         int CurrentY = this.tile.GetMinedY();
 	         int LastY = this.tile.GetLastY();
 	         int HoleSize = this.tile.GetSize();
 	         
 	         int YLeft = CurrentY - LastY;
 	         
-	         fontRendererObj.drawString(StatCollector.translateToLocal("words.power") + ": " + Power + "/" + tile.GetMaxPower(), x + 59, y + 17, 0x000000);
+	         fontRendererObj.drawString(StatCollector.translateToLocal("words.power") + ": " + Power + "/" + (int)tile.GetMaxPower(), x + 59, y + 17, 0x000000);
 	         fontRendererObj.drawString(StatCollector.translateToLocal("words.holesize") + ": " + HoleSize + "x" + HoleSize, x + 59, y + 27, 0x000000);
 	         if(YLeft > 0)
 	         fontRendererObj.drawString(StatCollector.translateToLocal("words.mining") + ": " + YLeft + " "+  StatCollector.translateToLocal("words.deeper") + ".", x + 59, y + 37, 0x000000);
